Block account deletion while user is referenced by tasks

+++ app/Http/Controllers/ProfileController.php
@@ -44,6 +44,11 @@ class ProfileController extends Controller
 
         $user = $request->user();
 
+        if ($user->createdTasks()->exists() || $user->assignedTasks()->exists()) {
+            flash(__('views.profile.flash.destroy.constraint'))->error();
+            return Redirect::route('profile.edit');
+        }
+
         Auth::logout();
 
         $user->delete();
